replaced size_t with int

+++ lab_11_01_01/inc/my_printf.h
@@ -1,1 +1,1 @@
-int my_snprintf(char *restrict s, size_t n, const char *restrict fmt, ...);
+int my_snprintf(char *restrict s, int n, const char *restrict fmt, ...);
